requirements.txt
Pillow>=9.0.0
numpy>=1.19.0
requests
